backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
nltk
textblob
transformers
torch
python-multipart
google-generativeai
python-dotenv